requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.0
dj-database-url==0.5.0
Django==2.2.3
django-elasticsearch-dsl==6.4.1
djangorestframework==3.10.1
drf-yasg==1.16.1
elasticsearch==6.4.0
elasticsearch-dsl==6.4.0
factory-boy==2.12.0
Faker==2.0.0
idna==2.8
importlib-metadata==0.18
inflection==0.3.1
ipython==7.6.1
ipython-genutils==0.2.0
itypes==1.1.0
jedi==0.14.1
Jinja2==2.10.1
MarkupSafe==1.1.1
more-itertools==7.2.0
packaging==19.0
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
pluggy==0.12.0
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
py==1.8.0
Pygments==2.4.2
pyparsing==2.4.0
pytest==5.0.1
pytest-django==3.5.1
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.1
requests==2.22.0
ruamel.yaml==0.15.100
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.2
traitlets==4.3.2
uritemplate==3.0.0
urllib3==1.25.3
wcwidth==0.1.7
zipp==0.5.2
